Refactor CodeGeneration
